+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -153,6 +153,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Cull Mode option for opaque materials and ShaderGraphs. 
 - Added scene view exposure override.
 - Added support for exposure curve remapping for min/max limits.
+- Added presets for ray traced reflections.
+- Added final image histogram debug view (both luminance and RGB).
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)
@@ -688,6 +690,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed an exception occuring when a camera doesn't have an HDAdditionalCameraData (1254383).
 - Fixed ray tracing with XR single-pass.
 - Fixed warning in HDAdditionalLightData OnValidate (cases 1250864, 1244578)
+- Fixed a bug related to denoising ray traced reflections.
+- Fixed nullref in the layered lit material inspector.
 
 ### Changed
 - Improve MIP selection for decals on Transparents
@@ -832,6 +836,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changing the default SSGI radius for the all configurations.
 - Changed the default parameters for quality RTGI to match expected behavior.
 - Add color clear pass while rendering XR occlusion mesh to avoid leaks.
+- Only use one texture for ray traced reflection upscaling.
+- Adjust the upscale radius based on the roughness value.
+- DXR: Changed the way the filter size is decided for directional, point and spot shadows.
+- Changed extensions of shader CAS include files.
 
 ## [7.1.1] - 2019-09-05